fix(update): Rename models
